requirements.txt
fastapi
uvicorn
streamlit
feedparser
trafilatura
sentence-transformers
faiss-cpu
numpy
requests
pydantic
